+++ bim_gw/utils/text_composer/composer.py
@@ -86,7 +86,7 @@ composer = Composer(script_structures, writers, variants, modifiers)
 
 if __name__ == '__main__':
 
-    args = get_args(debug=int(os.getenv("DEBUG", 0)))
+    args = get_args(debug=bool(os.getenv("DEBUG", False)))
     all_choices = inspect_all_choices(composer)
     dataset_location = Path(args.simple_shapes_path)
     for split in ["train", "val", "test"]: